requirements.txt
fastapi==0.103.1
uvicorn==0.22.0
pydantic==2.5.2
pillow==10.4.0
reportlab==4.2.2
